refactor: IntersectionObserverScroll 공통 컴포넌트 디렉토리로 분리 #91
