+++ frontend/src/components/IntersectionObserverScroll/IntersectionObserverScroll.tsx
