+++ frontend/src/pages/RecommendTodoPage/components/RecommendTodoContainer/index.tsx
@@ -11,12 +11,12 @@ import FilterForm from '../FilterForm';
 import BottomSheet from '@/components/BottomSheet';
 import IconButton from '@/components/Button/IconButton';
 import Icon from '@/components/Icon';
+import IntersectionObserverScroll from '@/components/IntersectionObserverScroll/IntersectionObserverScroll';
 import { MAX_TODO_ITEM_LENGTH, TODO_TYPE } from '@/constants/config';
 import { TODO_TOAST_MESSAGE } from '@/constants/message';
 import useQueryParamsDate from '@/hooks/useQueryParamsDate';
 import useToast from '@/hooks/useToast';
 import useTodoListQuery from '@/hooks/useTodoListQuery';
-import IntersectionObserverScroll from '@/pages/EditPage/components/IntersectionObserverScroll/IntersectionObserverScroll';
 import TodoEditItem from '@/pages/EditPage/components/TodoEditItem';
 import useAddTodoFromArchivedMutation from '@/pages/EditPage/hooks/useAddTodoFromArchivedMutation';
 import { CategoryType, DifficultyType } from '@/types/filter';
